integrate into main

+++ edge_random.cpp
@@ -5,7 +5,7 @@
 
 
 
-std::vector<int> EdgeRandom(std::vector<std::vector<int>> Matrix, std::vector<int> cycles, unsigned timeNs)
+std::vector<int> EdgeRandom(std::vector<std::vector<int>> Matrix, std::vector<int> cycles, uint64_t timeNs)
 {
     auto CycleA = std::vector<int>(cycles.begin(), cycles.begin() + cycles.size() / 2);
 	auto CycleB = std::vector<int>(cycles.begin() + cycles.size() / 2, cycles.end());
